requirements.txt
finance-datareader==0.9.90
python-dotenv==0.21.0
PyMySQL==1.1.0
pandas==2.0.3
beautifulsoup4===4.12.2
plotly===5.9.0